// prismaClient/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const PropertyType = {
  APARTMENT: 'APARTMENT',
  CONDO: 'CONDO',
  HOUSE: 'HOUSE',
  STUDIO: 'STUDIO',
  COTTAGE_OR_CABIN: 'COTTAGE_OR_CABIN',
  CHALET: 'CHALET'
} as const

export type PropertyType = (typeof PropertyType)[keyof typeof PropertyType]